Abstract rmcp stdio transport runtime
Move local stdio process startup behind a runtime trait while preserving the existing local MCP stdio behavior.
